fix(timechart): harden grid descriptors and reject unsafe arithmetic
Revalidate exact boundaries against immutable search bounds, reject incompatible source-presence flags before sparse publication, and retain the existing bin parser surface. Remove superseded fixed-grid planning code and bound rejected automatic calendar candidates before allocation. Focused parser/planner tests and existing ClickHouse/executor timechart suites pass. No visual changes.

+++ internal/clickhouse/compiler.go
@@ -658,7 +658,7 @@ type TimechartOutput struct {
 
 func validTimechartOutputSpanContract(output *TimechartOutput) bool {
 	return output != nil &&
-		((output.ExactGrid && len(output.Boundaries) == int(output.BucketCount)+1) || (output.Calendar && output.Span == 0) ||
+		((output.ExactGrid && uint64(len(output.Boundaries)) == output.BucketCount+1) || (output.Calendar && output.Span == 0) ||
 			(!output.Calendar && output.Span > 0))
 }
